pint float

+++ YouTank/Jogador.cpp
@@ -11,7 +11,7 @@ void Jogador::setDirecao_y(float dir_y)
 	direcao.y = dir_y;
 }
 
-Jogador::Jogador(int vida, int dano, int id, int barraVida_y, int player):
+Jogador::Jogador(int vida, int dano, int id, float barraVida_y, int player):
 	Personagem(vida, dano, id),
 	colisaoBot(false),
 	barraVida(static_cast<Personagem*>(this), 25.f, barraVida_y, 210.f, 25.f, player)
